Manfu Agentic/requirements.txt
langchain==0.2.16
langchain-openai==0.1.25
langchain-core==0.2.38
openai==1.45.0
streamlit==1.38.0
plotly==5.24.1
pandas==2.2.2
python-dotenv==1.0.1
